requirements.txt
fastapi==0.111.0
pydantic==2.7.4
PyJWT==2.8.0
python-dotenv==1.0.1
Requests==2.32.3
uvicorn==0.30.1
hypercorn==0.14.4
python-docx==1.1.2
docx2pdf==0.1.8
typing_extensions==4.9.0
pypandoc==1.13
python-multipart==0.0.9
mysql-connector-python==8.3.0
PyJWT==2.8.0
python-dotenv==1.0.1
SQLAlchemy==2.0.25
urllib3==2.2.1
Requests==2.32.3
pydantic[email]
